Entity Framework and Repository the  Pattern
Created Instance of DataContext on the Home Controller on
eCommerce.WebUI and can access all Models (Product - Customer Orderitem
etc)
But this is NOT GOOD! because we don't get the abstraction or separation
we want. We DO NOT want to use DbContext to access EntityFramework
directly because everything we want to do will be done thru Controller
accessing the EntityFramework directly by the DbContext. ==> SOLUTION:
Abstract the whole Data Context away by using REPOSITORIES