Using Interfaces - Turning RepositoryBase.cs into an Interface
Using Interfaces - Turning RepositoryBase.cs into an Interface
PROBLEM - Our program is breaking SOLID principles. This is becz for
every Controller that we want to use a repository in we have to define
the Repository and pass to it a DataContext. Problem now comes when we
want to change the concrete implementation of one of the repositories,
for example instead of using SQL server it wants to use Oracle or get
data from an extrernal source from the internet.
SOLUTION: Dependancy Injection, which uses Interfaces.  Once we
implement an interface.  An Interface is a means of specifying the
inputs and outputs of a perticular class without providing the actual
implementation. Benefit is once created, any implementation of that
Interface just needs to Inherit the Interface

IRepositoryBase.cs is the Inteface class created from RepositoryBase.cs,
which now inherits from  Interface class - IRepositoryBase.cs