Dependancy Injection with Unity
Injected the Repository into the Controller rather than instantiating it
directly
First creat a contrcuctor for the controller, and it's part of that
constructor that we want to put the implementaion details of the
context.
NOTE: Constructor doesn't work if you create a constructor with a
parameter in it.  A Dependancy Injector will be used instead, Unity is
the one i used, downloaded as a NuGET package.
UnityConfig.cs is automatically created under APP_Start folder after
downloading it. Added following line for Customer Repository:

container.RegisterType<IRepositoryBase<Customer>, CustomerRepository>();