Module Injection - Basic Voucher System part 1
Where we take a process that has common input and output & create
multiple implementations of that process. It uses Interfaces. Each
implementation is a self-contained piece of code that is usually a
single class library project, however doesn't necessarilly have to be!

In addition to the Customer & Product Repositoris below (for database
conn):

class CustomerRepository : RepositoryBase<Product>
class ProductrRepository : RepositoryBase<Customer>

I Created a VoucherRepository that is inhering from
RepositoryBase<Voucher>

class VoucherRepository: RepositoryBase<Voucher>

Each implementation of the process will be referenced by adding through
add references like we normally would, but by direcly referencing the
compiled class. This means we will have to know the full name of the
compiled class & we can get this from the database or a config file.. It
is also possible to do file discovery but that is a bit complex, will do
it later!
Each class created that implement the process will need to adhere &
implement a common Interface. Remember Interfaces allow to define input
& output but have a different implementation because the calling code
references the Interface and not the concete class, you get the ultimate
inflexibility!

Created Voucher class then refactored it and created an Interface called
IVoucher, wgich i moved into it's own separate file!

Created VoucherType class then refactored it and created an Interface
called IVoucherType, which i moved into it's own separate file!